<feat> Set up router
- npm installed vue-router
- Wrapped original App.vue from Kevan into KevanTest component
- Created routing table
- Created RouteView as the App.vue landing page for now.
- General clean up of KevanTest imports
