+++ Pawsitive/package.json
@@ -13,7 +13,8 @@
     "lint": "eslint . --fix"
   },
   "dependencies": {
-    "vue": "^3.5.18"
+    "vue": "^3.5.18",
+    "vue-router": "^4.0.13"
   },
   "devDependencies": {
     "@eslint/js": "^9.31.0",
