+++ Pawsitive/src/components/RouterTest.vue
@@ -1,0 +1,28 @@
+<script setup>
+
+</script>
+
+<script>
+    export default { 
+        data() {
+            return {
+                
+            }
+        }
+    }
+</script>
+
+<template>   
+
+    <RouterLink to="/KevanTest/">Kevan's Test</RouterLink> <br>
+    <RouterLink to="/TheWelcome/">The Welcome</RouterLink> <br>
+    <RouterLink to="/HelloWorld/">Hello World</RouterLink> <br>
+   
+</template>
+
+<style scope>
+    a {
+        margin: 5px;
+        display: block;
+    }
+</style>
